# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-commerce Database Optimization API",
        "description": "API for managing consolidated e-commerce orders with optimized analytics performance",
        "contact": {
            "name": "Lawal Owolabi Yusuf",
            "email": "lawalyusuf@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Local development server"
        }
    ],
    "paths": {
        "/consolidated-orders": {
            "get": {
                "tags": [
                    "Analytics"
                ],
                "summary": "Get analytics data",
                "description": "Retrieve comprehensive analytics data from consolidated orders with optional filtering",
                "operationId": "ccc37c4f05c6ac4960ef1553d9587127",
                "parameters": [
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Start date filter (Y-m-d format)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2024-01-01"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "End date filter (Y-m-d format)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2024-12-31"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Order status filter",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "delivered"
                        }
                    },
                    {
                        "name": "customer_id",
                        "in": "query",
                        "description": "Customer ID filter",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Analytics data retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "total_revenue": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 1807.45
                                                },
                                                "total_orders": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "total_items": {
                                                    "type": "integer",
                                                    "example": 6
                                                },
                                                "avg_order_value": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 602.48
                                                },
                                                "top_products": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "revenue_by_month": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "filters": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/consolidated-orders/populate": {
            "post": {
                "tags": [
                    "Data Management"
                ],
                "summary": "Populate consolidated orders",
                "description": "Trigger the population/refresh of consolidated orders table with optimized batch processing",
                "operationId": "1fbcd2e26ed21d044801d1f55bb8111b",
                "responses": {
                    "200": {
                        "description": "Data populated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully processed 345 records"
                                        },
                                        "records_processed": {
                                            "type": "integer",
                                            "example": 345
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error during population",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error populating consolidated orders: Database connection failed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/consolidated-orders/export": {
            "get": {
                "tags": [
                    "Export/Import"
                ],
                "summary": "Export to Excel",
                "description": "Download consolidated orders data as Excel file with optional filtering",
                "operationId": "d985ec06616e12bacdfe5feb5e7778a1",
                "parameters": [
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Start date filter (Y-m-d format)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2024-01-01"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "End date filter (Y-m-d format)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2024-12-31"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Order status filter",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "delivered"
                        }
                    },
                    {
                        "name": "customer_id",
                        "in": "query",
                        "description": "Customer ID filter",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Excel file download",
                        "content": {
                            "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Export error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error exporting data: File generation failed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/consolidated-orders/import": {
            "post": {
                "tags": [
                    "Export/Import"
                ],
                "summary": "Import from Excel",
                "description": "Upload and import consolidated orders data from Excel file with validation",
                "operationId": "1bc1df48f6f85d2db9c5c9883f1694b6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "file": {
                                        "description": "Excel file (.xlsx, .xls, .csv)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Import completed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Import completed successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        },
                                        "errors": {
                                            "properties": {
                                                "file": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The file field is required."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Import error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error importing data: File format not supported"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Analytics",
            "description": "Analytics"
        },
        {
            "name": "Data Management",
            "description": "Data Management"
        },
        {
            "name": "Export/Import",
            "description": "Export/Import"
        }
    ]
}